@@ -3,6 +3,7 @@
 	"version": "0.0.0",
 	"private": true,
 	"bin": "bin/cdk.js",
+	"type": "module",
 	"scripts": {
 		"preinstall": "npx only-allow pnpm",
 		"build": "tsc",
@@ -11,33 +12,34 @@
 		"test:dev": "jest --watch",
 		"format": "prettier --write \"{lib,bin}/**/*.ts\"",
 		"cdk": "cdk",
-		"lint": "eslint lib/** bin/** --ext .ts --no-error-on-unmatched-pattern",
+		"lint": "eslint lib/**/*.ts bin/**",
 		"synth": "cdk synth --path-metadata false --version-reporting false",
 		"diff": "cdk diff --path-metadata false --version-reporting false"
 	},
 	"devDependencies": {
 		"@guardian/cdk": "^58.1.2",
-		"@guardian/eslint-config-typescript": "^9.0.3",
-		"@guardian/prettier": "^8.0.0",
-		"@guardian/tsconfig": "^0.2.0",
+		"@guardian/eslint-config-typescript": "^10.0.1",
+		"@guardian/prettier": "^8.0.1",
+		"@guardian/tsconfig": "^1.0.0",
 		"@types/jest": "^29.5.12",
-		"@types/node": "20.14.8",
-		"aws-cdk": "2.147.0",
-		"aws-cdk-lib": "2.147.0",
+		"@types/node": "20.14.9",
+		"aws-cdk": "2.147.2",
+		"aws-cdk-lib": "2.147.2",
 		"constructs": "10.3.0",
-		"eslint": "^8.57.0",
+		"eslint": "^9.6.0",
 		"jest": "^29.7.0",
 		"ts-jest": "^29.1.5",
 		"ts-node": "^10.9.2",
-		"tslib": "^2.6.2",
+		"tslib": "^2.6.3",
 		"typescript": "~5.5.2"
 	},
 	"dependencies": {
-		"npm": "^10.4.0",
-		"source-map-support": "^0.5.20"
+		"npm": "^10.8.1",
+		"source-map-support": "^0.5.21"
 	},
 	"resolutions": {
 		"wrap-ansi": "7.0.0"
 	},
 	"prettier": "@guardian/prettier"
 }
+
