@@ -1,48 +1,52 @@
 {
   "name": "monitoring",
   "version": "1.0.0",
+  "type": "module",
   "main": "index.js",
   "license": "MIT",
   "scripts": {
     "test": "jest",
     "test:dev": "jest --watch",
-    "lint": "eslint src/** localRun.ts --ext .ts --no-error-on-unmatched-pattern",
-    "start": "ts-node localRun.ts",
-    "remote": "ts-node remoteRun.ts",
-    "build": "ncc build src/index.ts -o dist -e aws-sdk",
-    "validate": "npm-run-all test lint build"
+    "lint": "eslint src/** localRun.ts",
+    "start": "ts-node --esm localRun.ts",
+    "remote": "ts-node --esm remoteRun.ts",
+    "validate": "npm-run-all test lint"
   },
   "dependencies": {
-    "@aws-sdk/client-cloudwatch": "^3.592.0",
-    "commander": "^11.1.0",
+    "@aws-sdk/client-cloudwatch": "^3.606.0",
+    "commander": "^12.1.0",
     "dotenv": "^16.4.5",
-    "inquirer": "^8.2.6",
+    "inquirer": "^9.3.2",
     "playwright-aws-lambda": "^0.10.0",
-    "playwright-core": "^1.44.1",
+    "playwright-core": "^1.45.0",
     "tslib": "^2.6.3",
-    "uuid": "^9.0.1"
+    "uuid": "^10.0.0"
   },
   "devDependencies": {
-    "@aws-sdk/client-lambda": "^3.592.0",
+    "@aws-sdk/client-lambda": "^3.606.0",
     "@babel/parser": "^7.24.7",
     "@guardian/eslint-config-typescript": "^10.0.1",
-    "@guardian/prettier": "^7.0.0",
+    "@guardian/prettier": "^8.0.1",
     "@guardian/tsconfig": "^1.0.0",
     "@tsconfig/node18": "^18.2.4",
-    "@types/aws-lambda": "^8.10.138",
-    "@types/inquirer": "^8.2.10",
+    "@types/aws-lambda": "^8.10.140",
+    "@types/commander": "^2.12.2",
+    "@types/inquirer": "^9.0.7",
     "@types/jest": "^29.5.12",
-    "@types/node": "^20.14.2",
+    "@types/node": "^20.14.9",
     "@vercel/ncc": "^0.38.1",
     "aws-lambda": "^1.0.7",
-    "eslint": "^8.57.0",
+    "eslint": "^9.6.0",
+    "eslint-plugin-compat": "^5.0.0",
+    "globals": "^15.7.0",
     "jest": "^29.7.0",
     "markdown-toc": "^1.2.0",
     "npm-run-all": "^4.1.5",
-    "ts-jest": "^29.1.4",
+    "ts-jest": "^29.1.5",
     "ts-node": "^10.9.2",
     "ts-node-dev": "^2.0.0",
-    "typescript": "5.4.5"
+    "typescript": "5.5.2",
+    "typescript-eslint": "^7.14.1"
   },
   "prettier": "@guardian/prettier",
   "overrides": {
